C++ version of lenet
Adding a templated C++ version of lenet, so that we can test several
float format.
To than aim, we use the floatx header only floating point  implementation.
As of now, only C float are supported, because the arrays containing
images and parameters are of this type.
Images are converted, but parameters still need to be converted.